requirements/requirements.txt
# Core requirements
requests==2.31.0
beautifulsoup4==4.12.2
speedtest-cli==2.1.3
urllib3==2.0.7

# Testing (lightweight - no problematic dependencies)
pytest==7.4.0
pytest-html==4.0.2
pytest-xdist==3.3.1

# CI/CD
python-dotenv==1.0.0
setuptools==68.2.2